+++ get_x_navigation_generator/example/lib/navigator.dart
@@ -1,4 +1,5 @@
 import 'package:example/base_page.dart';
+import 'package:example/main.dart';
 import 'package:get_x_navigation_generator_annotations/get_x_navigation_generator_annotations.dart';
 
 import 'navigator.get_x_navigator.dart';
@@ -17,5 +18,16 @@ class MainNavigator extends GeneratedNavigator {
 
   static String get initialRoute => RouteNames.firstPage;
 
-  static final pages = GeneratedNavigator.pages;
+  // You can add additional pages here, for example a default page with an empty route
+  static final pages = [
+    BasePage<dynamic>(
+      name: '/',
+      page: () => const FirstPage(),
+    ),
+    ...GeneratedNavigator.pages,
+  ];
+
+  // This is purely for the example app to show how to use a navigator id, most likely you will hardcode this
+  // in the page that uses the navigator or have some logic to determine which navigator id to use
+  static int? lastNavigatorId;
 }